requirements.txt
python-consul>=0.3.3
boto>=2.36